+++ mcycle/_constants.pyx
@@ -97,6 +97,10 @@ cdef public unsigned char FLOWSENSE_UNDEFINED = 0
 cdef public unsigned char COUNTERFLOW = 1
 cdef public unsigned char PARALLELFLOW = 2
 cdef public unsigned char CROSSFLOW = 3
+# Constraints
+cdef public unsigned char NO_CONSTRAINT = 0
+cdef public unsigned char CONSTANT_P = 1
+cdef public unsigned char CONSTANT_V = 2
 # MCycle
 cdef public str SOURCE_URL = 'https://github.com/momargoh/MCycle'
 cdef public str DOCS_URL = 'https://mcycle.momarhughes.com'