+++ mcycle/components/coolers/__init__.py
@@ -1,1 +1,1 @@
-from .clr_basic import ClrBasic, ClrBasicConstP, ClrBasicConstV
+from .clr_basic import ClrBasic