+++ mcycle/components/heaters/__init__.py
@@ -1,1 +1,1 @@
-from .htr_basic import HtrBasic, HtrBasicConstP, HtrBasicConstV
+from .htr_basic import HtrBasic